pyproject.toml
[tool.poetry]
name = "deephaven-black"
version = "0.1.0.dev0"
description = "black formatting plugin for Deephaven"
authors = ["Devin Smith"]
readme = "README.md"
packages = [{include = "deephaven_black"}]

[tool.poetry.dependencies]
python = "^3.10"
deephaven-plugin = "^0.6.0"
black = "^23.12.0"

[tool.poetry.plugins."deephaven.plugin"]
registration_cls = "deephaven_black._register:BlackRegistration"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
